Refactored tests into separate files
